server/requirements.txt
Flask>=2.0.3
pymongo>=4.3.3
python-dotenv>=1.0.0
dnspython>=2.3.0
flask-cors>=3.0.10
